Consolidated proposal action logic.
