+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalSingle/components/ProposalStatusAndInfo.tsx
@@ -1,18 +1,14 @@
 import {
   AccountCircleOutlined,
   ArrowOutwardRounded,
-  CancelOutlined,
   HourglassTopRounded,
-  Key,
   Redo,
   RotateRightOutlined,
-  Send,
   Tag,
   ThumbUpOutlined,
 } from '@mui/icons-material'
 import clsx from 'clsx'
-import { useCallback, useEffect, useState } from 'react'
-import toast from 'react-hot-toast'
+import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import TimeAgo from 'react-timeago'
 import { useRecoilValue } from 'recoil'
@@ -22,7 +18,6 @@ import {
   CopyToClipboardUnderline,
   IconButtonLink,
   Logo,
-  ProposalCrossChainRelayStatus,
   ProposalStatusAndInfoProps,
   ProposalStatusAndInfo as StatelessProposalStatusAndInfo,
   Tooltip,
@@ -44,7 +39,6 @@ import {
   formatDateTimeTz,
   formatPercentOf100,
   getProposalStatusKey,
-  processError,
 } from '@dao-dao/utils'
 
 import { ButtonLink, SuspenseLoader } from '../../../../components'
@@ -53,11 +47,11 @@ import {
   CwProposalSingleV1Hooks,
   DaoProposalSingleV2Hooks,
   useAwaitNextBlock,
-  useMembership,
   useProposalPolytoneState,
   useProposalVetoState,
   useWallet,
 } from '../../../../hooks'
+import { useProposalActionState } from '../../../../hooks/useProposalActionState'
 import { useProposalModuleAdapterOptions } from '../../../react'
 import {
   useCastVote,
@@ -131,10 +125,7 @@ const InnerProposalStatusAndInfo = ({
   const { coreAddress } = useDaoInfoContext()
   const { getDaoProposalPath } = useDaoNavHelpers()
   const { proposalModule, proposalNumber } = useProposalModuleAdapterOptions()
-  const { isWalletConnected, address: walletAddress = '' } = useWallet()
-  const { isMember = false } = useMembership({
-    coreAddress,
-  })
+  const { address: walletAddress = '' } = useWallet()
 
   const config = useRecoilValue(
     DaoProposalSingleCommonSelectors.configSelector({
@@ -199,13 +190,6 @@ const InnerProposalStatusAndInfo = ({
     sender: walletAddress,
   })
 
-  const [actionLoading, setActionLoading] = useState(false)
-  // On proposal status update, stop loading. This ensures the action button
-  // doesn't stop loading too early, before the status has refreshed.
-  useEffect(() => {
-    setActionLoading(false)
-  }, [proposal.status])
-
   const polytoneState = useProposalPolytoneState({
     msgs: proposal.msgs,
     status: proposal.status,
@@ -215,53 +199,15 @@ const InnerProposalStatusAndInfo = ({
     openSelfRelayExecute,
     loadingTxHash: loadingExecutionTxHash,
   })
-
-  const onExecute = useCallback(async () => {
-    if (!isWalletConnected) {
-      return
-    }
-
-    setActionLoading(true)
-    try {
-      await executeProposal({
-        proposalId: proposalNumber,
-      })
-
-      await onExecuteSuccess()
-    } catch (err) {
-      console.error(err)
-      toast.error(processError(err))
-
-      // Stop loading if errored.
-      setActionLoading(false)
-    }
-
-    // Loading will stop on success when status refreshes.
-  }, [isWalletConnected, executeProposal, proposalNumber, onExecuteSuccess])
-
-  const onClose = useCallback(async () => {
-    if (!isWalletConnected) {
-      return
-    }
-
-    setActionLoading(true)
-
-    try {
-      await closeProposal({
-        proposalId: proposalNumber,
-      })
-
-      await onCloseSuccess()
-    } catch (err) {
-      console.error(err)
-      toast.error(processError(err))
-
-      // Stop loading if errored.
-      setActionLoading(false)
-    }
-
-    // Loading will stop on success when status refreshes.
-  }, [isWalletConnected, closeProposal, proposalNumber, onCloseSuccess])
+  const { action, footer } = useProposalActionState({
+    statusKey,
+    polytoneState,
+    loadingExecutionTxHash,
+    executeProposal,
+    closeProposal,
+    onExecuteSuccess,
+    onCloseSuccess,
+  })
 
   const awaitNextBlock = useAwaitNextBlock()
   // Refresh proposal and list of proposals (for list status) once voting ends.
@@ -527,49 +473,8 @@ const InnerProposalStatusAndInfo = ({
   return (
     <StatelessProposalStatusAndInfo
       {...props}
-      action={
-        statusKey === ProposalStatusEnum.Passed &&
-        // Show if anyone can execute OR if the wallet is a member, once
-        // polytone messages that need relaying are done loading.
-        (!config.only_members_execute || isMember) &&
-        !polytoneState.loading
-          ? {
-              label: t('button.execute'),
-              Icon: Key,
-              loading: actionLoading,
-              doAction: polytoneState.data.needsSelfRelay
-                ? polytoneState.data.openPolytoneRelay
-                : onExecute,
-            }
-          : statusKey === ProposalStatusEnum.Rejected
-          ? {
-              label: t('button.close'),
-              Icon: CancelOutlined,
-              loading: actionLoading,
-              doAction: onClose,
-            }
-          : // If executed and has polytone messages that need relaying...
-          statusKey === ProposalStatusEnum.Executed &&
-            !polytoneState.loading &&
-            polytoneState.data.needsSelfRelay &&
-            !loadingExecutionTxHash.loading &&
-            loadingExecutionTxHash.data
-          ? {
-              label: t('button.relay'),
-              Icon: Send,
-              loading: actionLoading,
-              doAction: polytoneState.data.openPolytoneRelay,
-              description: t('error.polytoneExecutedNoRelay'),
-            }
-          : undefined
-      }
-      footer={
-        !polytoneState.loading &&
-        statusKey === ProposalStatusEnum.Executed &&
-        polytoneState.data.hasPolytoneMessages && (
-          <ProposalCrossChainRelayStatus state={polytoneState.data} />
-        )
-      }
+      action={action}
+      footer={footer}
       info={info}
       status={status}
       vetoOrEarlyExecute={vetoOrEarlyExecute}
